move sensitive wps to only be able to be called from localhost, not from webmap nginx.
