+++ onomap-webmap/default.conf
@@ -20,6 +20,18 @@ server {
         xslt_stylesheet /ngx-niceindex/niceindex.xslt dir='$uri';
     }
 
+    location /dump {
+        alias /onomap/workspace/onomap_public_dump;
+        autoindex on;
+        autoindex_format xml;
+        xslt_stylesheet /ngx-niceindex/niceindex.xslt dir='$uri';
+    }
+
+    location /dump/.niceindex/ {
+        alias   /ngx-niceindex/;
+        index  index.html index.htm;
+    }
+
     location /cloudbeaver {
         proxy_pass http://cloudbeaver:8978;
         proxy_set_header Host $host;
