+++ onomap/src/main/groovy/org/noise_planet/onomap/sensitive/nc_process.groovy
@@ -24,7 +24,7 @@
  *  14-20 Boulevard Newton Cite Descartes, Champs sur Marne F-77447 Marne la Vallee Cedex 2 FRANCE
  *   or write to scientific.computing@univ-eiffel.fr
  */
-package org.noise_planet.onomap
+package org.noise_planet.onomap.sensitive
 
 
 import groovy.sql.Sql
