+++ onomap/src/test/groovy/org/noise_planet/onomap/TestNoiseCaptureGetPartyStats.groovy
@@ -29,6 +29,7 @@ package org.noise_planet.onomap
 import groovy.sql.Sql
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import org.noise_planet.onomap.sensitive.nc_parse
 
 import static org.junit.jupiter.api.Assertions.assertEquals
 import static org.junit.jupiter.api.Assertions.assertNotNull
