+++ onomap/src/test/groovy/org/noise_planet/onomap/TestNoiseCaptureParse.groovy
@@ -30,6 +30,7 @@ import groovy.sql.Sql
 import groovy.transform.CompileStatic
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import org.noise_planet.onomap.sensitive.nc_parse
 
 import java.sql.ResultSet
 import java.sql.Statement
